functions.js
fixed repeated notifications when refreshing the page
