+++ sources/js/constants.js
@@ -1,0 +1,7 @@
+// --- sources/js/constants.js ---
+export const MILLISECONDS_PER_SECOND = 1000;
+export const MILLISECONDS_PER_MINUTE = 60 * MILLISECONDS_PER_SECOND;
+export const MILLISECONDS_PER_HOUR = 60 * MILLISECONDS_PER_MINUTE;
+export const MILLISECONDS_PER_DAY = 24 * MILLISECONDS_PER_HOUR;
+
+export const SERVER_TIMEZONE = "America/Toronto";  // used for determining Daylight Saving Time
